interoperability with systems who disregard resources
the slack gateway story:
 - you log in
 - everybody is subscribed to everybody else
 - they use only bare (foo@bar.com) jabber ids, no resources
 - presence information, messages, etc. originate from the bare jid
 - deal with that by introducing the fake resource ""
   (to store presence information, otr context)
 - when sending a message, send it to the bare jid (if resource = "")
 --> OTR via slack works (thx to drup for testing)

this is in violation of XMPP RFC, but since it is deployed, there's
a stronger argument to deal with it.

as a side note, this should fix issues between a and b if s2s connection does
not transport presence information, but messages (see the removed comment in
user.ml).